requirements.txt
odfpy>=1.4.1
Pillow>=10.0.0
requests>=2.31.0
